Rename plain-text to html-to-text.
