+++ src/content/utils/html-to-text.spec.js
@@ -1,6 +1,6 @@
 import {expect} from 'chai';
 
-import {htmlToText} from './plain-text.js';
+import htmlToText from './html-to-text.js';
 
 describe('htmlToText', () => {
     it('should convert html to text', () => {
